+++ orte/mca/ess/alps/ess_alps_module.c
@@ -12,7 +12,7 @@
  * Copyright (c) 2011      Cisco Systems, Inc.  All rights reserved.
  * Copyright (c) 2011-2013 Los Alamos National Security, LLC.
  *                         All rights reserved.
- * Copyright (c) 2017      Intel, Inc. All rights reserved.
+ * Copyright (c) 2017-2018 Intel, Inc.  All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -30,8 +30,6 @@
 #include "orte/util/proc_info.h"
 #include "orte/mca/errmgr/base/base.h"
 #include "orte/util/name_fns.h"
-#include "orte/util/nidmap.h"
-#include "orte/util/regex.h"
 #include "orte/runtime/orte_globals.h"
 
 #include "orte/mca/ess/ess.h"
