+++ orte/mca/odls/base/odls_base_default_fns.c
@@ -14,7 +14,7 @@
  * Copyright (c) 2011-2013 Los Alamos National Security, LLC.
  *                         All rights reserved.
  * Copyright (c) 2011-2017 Cisco Systems, Inc.  All rights reserved
- * Copyright (c) 2013-2017 Intel, Inc.  All rights reserved.
+ * Copyright (c) 2013-2018 Intel, Inc.  All rights reserved.
  * Copyright (c) 2014-2017 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * Copyright (c) 2017      Mellanox Technologies Ltd. All rights reserved.
@@ -66,6 +66,7 @@
 #include "orte/mca/ess/base/base.h"
 #include "orte/mca/grpcomm/base/base.h"
 #include "orte/mca/plm/base/base.h"
+#include "orte/mca/regx/regx.h"
 #include "orte/mca/rml/base/rml_contact.h"
 #include "orte/mca/rmaps/rmaps_types.h"
 #include "orte/mca/rmaps/base/base.h"
@@ -78,10 +79,8 @@
 
 #include "orte/util/context_fns.h"
 #include "orte/util/name_fns.h"
-#include "orte/util/regex.h"
 #include "orte/util/session_dir.h"
 #include "orte/util/proc_info.h"
-#include "orte/util/nidmap.h"
 #include "orte/util/show_help.h"
 #include "orte/util/threads.h"
 #include "orte/runtime/orte_globals.h"
@@ -138,7 +137,7 @@ int orte_odls_base_default_get_add_procs_data(opal_buffer_t *buffer,
     /* if we couldn't provide the allocation regex on the orted
      * cmd line, then we need to provide all the info here */
     if (!orte_nidmap_communicated) {
-        if (ORTE_SUCCESS != (rc = orte_util_nidmap_create(orte_node_pool, &nidmap))) {
+        if (ORTE_SUCCESS != (rc = orte_regx.nidmap_create(orte_node_pool, &nidmap))) {
             ORTE_ERROR_LOG(rc);
             return rc;
         }
@@ -157,7 +156,7 @@ int orte_odls_base_default_get_add_procs_data(opal_buffer_t *buffer,
         orte_get_attribute(&jdata->attributes, ORTE_JOB_LAUNCHED_DAEMONS, NULL, OPAL_BOOL)) {
         flag = 1;
         opal_dss.pack(buffer, &flag, 1, OPAL_INT8);
-        if (ORTE_SUCCESS != (rc = orte_util_encode_nodemap(buffer))) {
+        if (ORTE_SUCCESS != (rc = orte_regx.encode_nodemap(buffer))) {
             ORTE_ERROR_LOG(rc);
             return rc;
         }
@@ -336,7 +335,7 @@ int orte_odls_base_default_get_add_procs_data(opal_buffer_t *buffer,
 
     if (!orte_get_attribute(&jdata->attributes, ORTE_JOB_FULLY_DESCRIBED, NULL, OPAL_BOOL)) {
         /* compute and pack the ppn regex */
-        if (ORTE_SUCCESS != (rc = orte_util_nidmap_generate_ppn(jdata, &nidmap))) {
+        if (ORTE_SUCCESS != (rc = orte_regx.generate_ppn(jdata, &nidmap))) {
             ORTE_ERROR_LOG(rc);
             return rc;
         }
@@ -514,7 +513,7 @@ int orte_odls_base_default_construct_child_list(opal_buffer_t *buffer,
             }
             /* populate the node array of the job map and the proc array of
              * the job object so we know how many procs are on each node */
-            if (ORTE_SUCCESS != (rc = orte_util_nidmap_parse_ppn(jdata, ppn))) {
+            if (ORTE_SUCCESS != (rc = orte_regx.parse_ppn(jdata, ppn))) {
                 ORTE_ERROR_LOG(rc);
                 free(ppn);
                 goto REPORT_ERROR;
