+++ orte/mca/regx/base/Makefile.am
@@ -1,0 +1,15 @@
+#
+# Copyright (c) 2015-2018 Intel, Inc.  All rights reserved.
+# $COPYRIGHT$
+#
+# Additional copyrights may follow
+#
+# $HEADER$
+#
+
+headers += \
+        base/base.h
+
+libmca_regx_la_SOURCES += \
+        base/regx_base_frame.c \
+        base/regx_base_select.c
