+++ orte/mca/regx/fwd/Makefile.am
@@ -1,0 +1,37 @@
+#
+# Copyright (c) 2016-2018 Intel, Inc.  All rights reserved.
+# Copyright (c) 2017      IBM Corporation.  All rights reserved.
+# $COPYRIGHT$
+#
+# Additional copyrights may follow
+#
+# $HEADER$
+#
+
+sources = \
+          regx_fwd_component.c \
+          regx_fwd.h \
+          regx_fwd.c
+
+# Make the output library in this directory, and name it either
+# mca_<type>_<name>.la (for DSO builds) or libmca_<type>_<name>.la
+# (for static builds).
+
+if MCA_BUILD_orte_regx_fwd_DSO
+component_noinst =
+component_install = mca_regx_fwd.la
+else
+component_noinst = libmca_regx_fwd.la
+component_install =
+endif
+
+mcacomponentdir = $(ortelibdir)
+mcacomponent_LTLIBRARIES = $(component_install)
+mca_regx_fwd_la_SOURCES = $(sources)
+mca_regx_fwd_la_LDFLAGS = -module -avoid-version
+mca_regx_fwd_la_LIBADD = $(top_builddir)/orte/lib@ORTE_LIB_PREFIX@open-rte.la
+
+noinst_LTLIBRARIES = $(component_noinst)
+libmca_regx_fwd_la_SOURCES = $(sources)
+libmca_regx_fwd_la_LDFLAGS = -module -avoid-version
+
