+++ orte/util/Makefile.am
@@ -11,7 +11,7 @@
 #                         All rights reserved.
 # Copyright (c) 2008      Sun Microsystems, Inc.  All rights reserved.
 # Copyright (c) 2014      Cisco Systems, Inc.  All rights reserved.
-# Copyright (c) 2014-2017 Intel, Inc.  All rights reserved.
+# Copyright (c) 2014-2018 Intel, Inc.  All rights reserved.
 # Copyright (c) 2016      Research Organization for Information Science
 #                         and Technology (RIST). All rights reserved.
 # $COPYRIGHT$
@@ -56,8 +56,6 @@ headers += \
         util/hostfile/hostfile_lex.h \
         util/dash_host/dash_host.h \
         util/comm/comm.h \
-        util/nidmap.h \
-        util/regex.h \
         util/attr.h \
         util/listener.h \
         util/compress.h \
@@ -77,8 +75,6 @@ lib@ORTE_LIB_PREFIX@open_rte_la_SOURCES += \
         util/hostfile/hostfile.c \
         util/dash_host/dash_host.c \
         util/comm/comm.c \
-        util/nidmap.c \
-        util/regex.c \
         util/attr.c \
         util/listener.c \
         util/compress.c
